requirements/base.txt
python-slugify==8.0.4
Pillow==10.3.0
argon2-cffi==23.1.0
whitenoise==6.6.0
redis==5.0.4
hiredis==2.3.2
celery==5.4.0
django-celery-beat==2.6.0
flower==2.0.1
uvicorn[standard]==0.29.0
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0
# Django
# ------------------------------------------------------------------------------
django==4.2
django-environ==0.10.0
django-model-utils==4.5.0
django-redis==5.4.0
django-mptt==0.16.0
django-filter==24.2
# Django REST Framework
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
# DRF-spectacular for api documentation
drf-spectacular==0.27.2
drf-nested-forms==1.1.8
# Other Dependencies
requests==2.28.2
